fix: 重启时释放内存
